# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Araldo API
  version: v1
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: |
    Schedule and publish posts to every platform from your product.

    **Authentication.** Send an API key as a bearer token:
    `Authorization: Bearer ald_test_…`. Test keys (`ald_test_`) only ever
    reach sandbox channels; live keys (`ald_live_`) reach real accounts.

    **Idempotency.** Every `POST` accepts an `Idempotency-Key` header. A
    retry with the same key and body returns the first response (with
    `Idempotent-Replayed: true`) for 24 hours.

    **Errors** are RFC 9457 problem details with a stable `code`, the
    `param` at fault, and every validation problem in `errors`.

    **Lists** page by cursor: `limit`, `starting_after`, `ending_before`,
    and `has_more` in the response.

    **Events and webhooks.** Every change of interest becomes an event,
    listed at `/v1/events` and delivered to your webhook endpoints, signed
    with `Araldo-Signature: t=<unix>,v1=<hex HMAC-SHA256 of "t.body">`.
servers:
  - url: /
security:
  - apiKey: []
...
paths:
...
  /v1/templates:
    get:
      summary: List templates
      tags: [Templates]
      parameters:
        - { name: brand, in: query, schema: { type: string } }
      responses:
        "200":
          description: Templates (without their bodies).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListBase"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Template" } }
        default: { $ref: "#/components/responses/Error" }
    post:
      summary: Create a template
      tags: [Templates]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/TemplateSource"
                - type: object
                  required: [brand, key]
                  properties:
                    brand: { type: string }
                    key: { type: string, example: featured-build }
                    name: { type: string }
                    approval: { $ref: "#/components/schemas/TemplateApproval" }
      responses:
        "201":
          description: The template with version 1.
          content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "An API key: ald_test_… or ald_live_…"
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 255 }
      description: Makes the request safe to retry for 24 hours.
...
  responses:
    Error:
      description: A problem.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string, example: too_long }
        param: { type: string }
        request_id: { type: string }
        doc_url: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              param: { type: string }
              message: { type: string }
              detail: { type: object, additionalProperties: true }
    ListBase:
      type: object
      required: [object, data, has_more, url]
      properties:
        object: { type: string, enum: [list] }
        data: { type: array, items: {} }
        has_more: { type: boolean }
        url: { type: string }
...
    TemplateSource:
      type: object
      properties:
        variables:
          type: object
          additionalProperties: true
          description: A JSON Schema (2020-12) for the data.
        examples: { type: array, items: { type: object, additionalProperties: true } }
        body: { type: string, example: "Featured build: {{.name}} {{.url}}" }
        overrides:
          type: object
          additionalProperties: { type: string }
          description: Per-platform bodies.
        fit:
          type: object
          additionalProperties: { type: string, enum: [error, truncate, thread] }
    TemplateApproval:
      type: string
      enum: [inherit, required, not_required]
      description: |
        Whether posts made from the template wait for approval: `inherit`
        follows the brand's policy (the default), `required` always waits,
        `not_required` never does.
    Template:
      type: object
      properties:
        id: { type: string }
        object: { type: string, enum: [template] }
        brand: { type: string }
        key: { type: string }
        name: { type: string }
        approval: { $ref: "#/components/schemas/TemplateApproval" }
        latest_version: { type: integer }
        latest:
          allOf:
            - $ref: "#/components/schemas/TemplateSource"
            - type: object
              properties:
                version: { type: integer }
                created_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...